pyproject.toml
[project]
name = "experimenting-with-pydoll"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.1.0",
    "pydoll-python>=1.7.0",
]
